requirements.txt
argparse
click 
ffmpeg-python==0.2.0
gTTS==2.3.1
json5 
numpy 
numpydoc 
openai==0.27.0
openai-whisper==20230314
PyAudio==0.2.13
pydub==0.25.1
pygame==2.1.2
queuelib==1.5.0
SpeechRecognition==3.9.0
threadpoolctl 
torch==2.0.0
ujson 
openai-whisper
